requirements.txt
asgiref==3.6.0
Django==3.2.18
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.5.0
